refactor : Reward controller 와 Users Reward controller를 분리함
- Reward controller와 Users Reward controller의 위치를 분리했습니다.
