+++ waldreg/application-layer/reward-controller/src/main/java/org/waldreg/controller/reward/tag/request/RewardTagRequest.java
@@ -1,4 +1,4 @@
-package org.waldreg.controller.reward.request;
+package org.waldreg.controller.reward.tag.request;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import javax.validation.constraints.NotNull;
