+++ waldreg/application-layer/reward-controller/src/main/java/org/waldreg/controller/reward/tag/response/RewardTagResponse.java
@@ -1,4 +1,4 @@
-package org.waldreg.controller.reward.response;
+package org.waldreg.controller.reward.tag.response;
 
 public final class RewardTagResponse{
